requirements.txt
appnope==0.1.2
asttokens==2.0.5
backcall==0.2.0
beautifulsoup4==4.11.1
bs4==0.0.1
cachetools==5.2.0
certifi==2020.12.5
chardet==4.0.0
click==8.1.3
cryptocompare==0.7.5
debugpy==1.5.1
decorator==5.1.1
entrypoints==0.4
executing==0.8.3
google-auth==2.14.1
google-auth-oauthlib==0.7.1
gsheetsdb==0.1.13.1
gspread==5.7.1
gspread-dataframe==3.3.0
idna==2.10
ipykernel==6.9.2
ipython==8.1.1
jedi==0.18.1
jupyter-client==7.1.2
jupyter-core==4.9.2
lxml==4.9.1
matplotlib-inline==0.1.3
mo-dots==4.22.21108
mo-future==3.147.20327
mo-imports==3.149.20327
mo-kwargs==4.22.21108
mo-logs==4.23.21108
mosspy==1.0.9
moz-sql-parser==4.40.21126
nest-asyncio==1.5.4
numerapi==2.12.4
numpy==1.23.3
oauthlib==3.2.2
pandas==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
plotly==5.11.0
prompt-toolkit==3.0.28
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.11.2
python-dateutil==2.8.2
pytz==2022.2.1
pyzmq==22.3.0
requests==2.25.1
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
soupsieve==2.3.2.post1
stack-data==0.2.0
tenacity==8.1.0
tornado==6.1
tqdm==4.64.1
traitlets==5.1.1
urllib3==1.26.4
wcwidth==0.2.5
